+++ docs/release_notes.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 This project uses [*towncrier*](https://towncrier.readthedocs.io/) and the changes for the upcoming release can be found in <https://github.com/fedora-infra/sqlalchemy-helpers/tree/develop/news/>.
 
 <!-- towncrier release notes start -->
+- Allow customization of the model base class.
+
 
 ## Version [0.12.1](https://github.com/fedora-infra/sqlalchemy-helpers/tree/0.12.1)